Use daps2docker in obs-docs deployment github action
Before switching to the daps2docker CLI, the github action used
an docker image which worked with an older version of DAPS (2.*.*
instead of 3.2.0).
This led to several little differences in the built docs, like
having `http` links instead of `https`. It is easier to rely on
the CLI tool in the github action over manually executing all
the necessary steps.